Split up widgets to make it a bit easier to work with.
